[X] Crash when setting mode setting to random in tournament mode [X] Weird world edtior behavior when selecting pipe minigame when exiting and viewing preview [X] Weird world editor behavior when selecting pipe minigame when paging down then up [X] Fixed bug where players with stored items would lose them in world mode if they finished a mode that didn't allow stored items [X] Added boxes minigame to world mode [X] Select which map is used for minigames and if map doesn't exist, choose default [X] Boss now collides with fireballs, hammers, boomerangs, bulletbills, attackzones [X] Fixed bug where player would shoot projectile the same time they were picking up an item.  This caused ice blasts to kill throw blocks on pickup. [X] Added freezing of enemies, shells, throw blocks/boxes [X] Tweak 1up-5up in star mode to give extra time except for 5up which gives a 1up [X] Give credit to killing another player with a falling donut [X] Button combo to drop cards in collection mode (turbo + turbo + turbo) [X] Add 2 secrets (No, I'm not telling you what they are :)) [X] Added new coin bonus to greed and coins mode [X] Added boss and boxes modes to world editor [X] Add extra coins to [?] boxes for greed and coin mode [X] Fixed AI to play boxes and shyguy modes ok [X] Started adding 2 more secrets

+++ _src/world.cpp
@@ -7,14 +7,14 @@
 #endif
 
 extern void ResetTourStops();
-extern TourStop * ParseTourStopLine(char * buffer, short iVersion[4], bool fIsWorld);
+extern TourStop * ParseTourStopLine(char * buffer, int iVersion[4], bool fIsWorld);
 extern void WriteTourStopLine(TourStop * ts, char * buffer, bool fIsWorld);
 extern WorldMap g_worldmap;
 extern bool LoadMenuSkin(short playerID, short skinID, short colorID, bool fLoadBothDirections);
 
 extern std::string stripPathAndExtension(const std::string &path);
 
-extern short g_iVersion[];
+extern int g_iVersion[];
 using std::queue;
 using std::map;
 
@@ -431,7 +431,7 @@ bool WorldMap::Load(short tilesize)
 
 	char buffer[1024];
 	short iReadType = 0;
-	short iVersion[4] = {0, 0, 0, 0};
+	int iVersion[4] = {0, 0, 0, 0};
 	short iMapTileReadRow = 0;
 	short iCurrentStage = 0;
 	short iCurrentWarp = 0;
